Extract more code from mock_gen to mock_utils
